refactor notifications

+++ ios/Runner/PlatformChannel.swift
@@ -18,7 +18,7 @@ class PlatformChannelManager {
     guard let arg = arg as? Dictionary<String, Any> else { throw PlatformError.BadArguments }
     guard let title = arg["title"] as? String else { throw PlatformError.BadArguments }
     guard let id = arg["id"] as? Int else { throw PlatformError.BadArguments }
-    guard let timeSince = arg["timeSince"] as? Int else {throw PlatformError.BadArguments }
+    guard let timeSince = arg["timeSince"] as? Double else {throw PlatformError.BadArguments }
     
     let res = notifications.scheduleNotification(text: title, timeSince: timeSince, id: id)
     return res
